fix: 🐛 修复了没有需要null删除的bug

+++ registry/src/main/java/com/ksyun/start/camp/dto/RegisteredServiceDTO.java
@@ -62,6 +62,9 @@ public class RegisteredServiceDTO {
                 break;
             }
         }
+        if (temp == null) {
+            return false;
+        }
         return goalService.remove(temp);
     }
